update to batch_starter.py
