+++ sds_data_manager/lambda_code/batch_starter.py
@@ -6,7 +6,6 @@ from datetime import datetime, timedelta
 
 import boto3
 import psycopg2
-from get_downstream_dependencies import get_downstream_dependencies
 
 # Setup the logging
 logging.basicConfig(level=logging.INFO)
@@ -388,6 +387,35 @@ def prepare_data(instruments_to_process):
     }
 
     return grouped_dict
+
+
+def get_downstream_dependencies(key):
+    """
+    Retrieves downstream dependencies of a given instrument.
+
+    Parameters
+    ----------
+    key : str
+        The key from the JSON file.
+
+    Returns
+    -------
+    dict
+        The value associated with the provided key in the JSON file.
+    """
+
+    # Construct the path to the JSON file
+    dir_path = os.path.dirname(os.path.realpath(__file__))
+    dependency_path = os.path.join(dir_path, "downstream_dependents.json")
+
+    with open(dependency_path) as file:
+        data = json.load(file)
+        value = data.get(key)
+
+        if value is None:
+            raise KeyError(f"Key '{key}' not found in the JSON file.")
+
+        return value
 
 
 def lambda_handler(event: dict, context):
